fix(code): babel-polyfill related bug
It’s not recommended to use babel-polyfill in a tool, see:
https://babeljs.io/docs/usage/polyfill/

babel-polyfill only allows a single instance to be loaded.

If you run webpack with `babel-node`, this causes the plugin to throw an error. This simple fix simply prevent `babel-polyfill` being loaded if it has previously been loaded into the environment.